requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.6.3
python-dotenv==1.0.1
motor==3.4.0
pymongo==4.6.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==3.2.0
email-validator
requests